refactor: 정팩메 추가
